\name{censo_2017_comunas}
\alias{censo_2017_comunas}
\docType{data}
\title{
Cifras de poblacion oficiales por comuna
}
\description{
Contiene los registros de poblacion desagregados por comuna, edad y sexo binario.
}
\usage{censo_2017_comunas}
\format{
Un data frame con 14.532 observaciones y cuatro columnas.
  \describe{
    \item{\code{codigo_comuna}}{vector de texto}
    \item{\code{edad}}{vector categorico}
    \item{\code{sexo}}{vector categorico}
    \item{\code{poblacion}}{vector numerico}
  }
}
\source{
Elaboracion propia con base en datos de Instituto Nacional de Estadisticas (INE).
}
\examples{
censo_2017_comunas
}
\keyword{datasets}
