\name{censo_2017_zonas}
\alias{censo_2017_zonas}
\docType{data}
\title{
Cifras de poblacion oficiales por zona
}
\description{
Contiene los registros de poblacion desagregados por zona y edad.
}
\usage{censo_2017_zonas}
\format{
Un data frame con 64.212 observaciones y tres columnas.
  \describe{
    \item{\code{geocodigo}}{vector de texto}
    \item{\code{edad}}{vector categorico}
    \item{\code{poblacion}}{vector numerico}
  }
}
\source{
Elaboracion propia con base en datos de Instituto Nacional de Estadisticas (INE).
}
\examples{
censo_2017_zonas
}
\keyword{datasets}
