\name{census_2017_communes}
\alias{census_2017_communes}
\docType{data}
\title{
2017 census data by commune
}
\description{
Dataset with official population statistics at commune level.
}
\usage{census_2017_communes}
\format{
Data frame

> str(census_2017_communes)
Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	346 obs. of  4 variables:
 $ commune_id: chr  "10101" "10102" "10103" "10104" ...
 $ pop_men   : num  120715 17104 2075 5992 8999 ...
 $ pop_women : num  124636 16881 1752 5950 9270 ...
 $ pop_total : num  245351 33985 3827 11942 18269 ...
}
\source{
Instituto Nacional de Estadisticas (INE)
}
\keyword{datasets}
