\name{codigos_territoriales}
\alias{codigos_territoriales}
\docType{data}
\title{
Codigos oficiales de comunas, provincias y regiones
}
\description{
Contiene los codigos oficiales de SUBDERE. Los nombres de las divisiones
administrativas fueron convertidos a ASCII para evitar problemas de
codificacion.
}
\usage{codigos_territoriales}
\format{
Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	346 obs. of  6 variables:
 $ codigo_comuna   : chr  "01101" "01107" "01401" "01402" ...
 $ nombre_comuna   : chr  "Iquique" "Alto Hospicio" "Pozo Almonte" "Camina" ...
 $ codigo_provincia: chr  "011" "011" "014" "014" ...
 $ nombre_provincia: chr  "Iquique" "Iquique" "Tamarugal" "Tamarugal" ...
 $ codigo_region   : chr  "01" "01" "01" "01" ...
 $ nombre_region   : chr  "Tarapaca" "Tarapaca" "Tarapaca" "Tarapaca" ...
}
\source{
Subsecretaria de Desarrollo Regional (SUBDERE)
}
\keyword{datasets}
