\name{divisiones_electorales}
\alias{divisiones_electorales}
\docType{data}
\title{
Divisiones electorales oficiales
}
\description{
Contiene las agrupaciones de las distintas comunas en circunscripciones
(eleccion de senadores) y distritos (eleccion de diputados) junto con la
cantidad de senadores y diputados que le corresponden a cada division
segun la legislacion vigente.
}
\usage{divisiones_electorales}
\format{
Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	346 obs. of  5 variables:
 $ codigo_comuna            : chr  "15101" "15102" "15202" "15201" ...
 $ codigo_circunscripcion   : chr  "01" "01" "01" "01" ...
 $ senadores_circunscripcion: int  2 2 2 2 2 2 2 2 2 2 ...
 $ codigo_distrito          : chr  "01" "01" "01" "01" ...
 $ diputados_distrito       : int  3 3 3 3 3 3 3 3 3 3 ...
}
\source{
Biblioteca del Congreso Nacional de Chile (BCN)
}
\keyword{datasets}
