\name{divisiones_salud}
\alias{divisiones_salud}
\docType{data}
\title{
Divisiones en servicios de salud oficiales
}
\description{
Contiene las agrupaciones de las distintas comunas en servicios de salud
segun la legislacion vigente.
}
\usage{divisiones_salud}
\format{
Un data frame con 346 observaciones y tres columnas.
  \describe{
    \item{\code{codigo_comuna}}{vector de texto}
    \item{\code{codigo_servicio_salud}}{vector de texto}
    \item{\code{nombre_servicio_salud}}{vector de texto}
  }
}
\source{
Elaboracion propia con base en las jurisdicciones que reporta el sitio web
de cada servicio de salud.
}
\examples{
divisiones_salud
}
\keyword{datasets}
