% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political-maps.R
\name{generar_distritos}
\alias{generar_distritos}
\title{Mapa de Chile a nivel de distritos}
\usage{
generar_distritos(mapa = chilemapas::mapa_comunas)
}
\arguments{
\item{mapa}{mapa a agregar, por defecto es todo el mapa nacional}
}
\value{
Un objeto de clase sf y data.frame.
}
\description{
Este mapa es calculado a partir
del mapa comunal para no recargar el volumen de datos del paquete.
}
\examples{
generar_distritos()
}
