% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative-maps.R
\name{generar_servicios_salud}
\alias{generar_servicios_salud}
\title{Mapa de Chile a nivel de servicios de salud}
\usage{
generar_servicios_salud(mapa = chilemapas::mapa_comunas)
}
\arguments{
\item{mapa}{mapa a agregar, por defecto es todo el mapa nacional}
}
\value{
Un objeto de clase sf y data.frame.
}
\description{
Este mapa es calculado a partir
del mapa comunal para no recargar el volumen de datos del paquete.
}
\examples{
generar_servicios_salud()
}
