\name{health_service_codes}
\alias{health_service_codes}
\docType{data}
\title{
Official health service codes (new regions)
}
\description{
Dataset with official administrative codes (communeand health service). The
special characters were converted to ASCII to avoid encoding problems. This
dataset considers 16 regions.
}
\usage{health_service_codes}
\format{
  A data frame with 346 observations on the following 3 variables.
  \describe{
    \item{\code{commune_id}}{a character vector}
    \item{\code{health_service_id}}{a character vector}
    \item{\code{health_service}}{a character vector}
  }
}
\source{
Infraestructura de Datos Geoespaciales Chile (IDE Chile)
}
\keyword{datasets}
