% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political-maps.R
\name{mapa_circunscripciones}
\alias{mapa_circunscripciones}
\title{Mapa de Chile a nivel de circunscripciones}
\usage{
mapa_circunscripciones()
}
\value{
un objeto de clase sf y data.frame, este mapa es calculado a partir
del mapa comunal para no recargar el volumen de datos del paquete
}
\description{
Mapa de Chile a nivel de circunscripciones
}
\examples{
# mapa_circunscripciones()
}
