% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative-maps.R
\name{mapa_provincias}
\alias{mapa_provincias}
\title{Mapa de Chile a nivel provincial}
\usage{
mapa_provincias()
}
\value{
un objeto de clase sf y data.frame, este mapa es calculado a partir
del mapa comunal para no recargar el volumen de datos del paquete
}
\description{
Mapa de Chile a nivel provincial
}
\examples{
# mapa_provincias()
}
