% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative-maps.R
\name{mapa_regiones}
\alias{mapa_regiones}
\title{Mapa de Chile a nivel regional}
\usage{
mapa_regiones()
}
\value{
un objeto de clase sf y data.frame, este mapa es calculado a partir
del mapa comunal para no recargar el volumen de datos del paquete
}
\description{
Mapa de Chile a nivel regional
}
\examples{
# mapa_regiones()
}
