\name{territorial_codes}
\alias{territorial_codes}
\docType{data}
\title{
Official administrative codes (new regions)
}
\description{
Dataset with official administrative codes (commune, province and region). The
special characters were converted to ASCII to avoid encoding problems. This
dataset considers 16 regions.
}
\usage{territorial_codes}
\format{
  A data frame with 346 observations on the following 6 variables.
  \describe{
    \item{\code{region_id}}{a character vector}
    \item{\code{region_name}}{a character vector}
    \item{\code{province_id}}{a character vector}
    \item{\code{province_name}}{a character vector}
    \item{\code{commune_id}}{a character vector}
    \item{\code{commune_name}}{a character vector}
  }
}
\source{
Subsecretaria de Desarrollo Regional (SUBDERE)
}
\keyword{datasets}
