\name{codigos_territoriales}
\alias{codigos_territoriales}
\docType{data}
\title{
Codigos oficiales de comunas, provincias y regiones
}
\description{
Contiene los codigos oficiales de SUBDERE. Los nombres de las divisiones
administrativas fueron convertidos a ASCII para evitar problemas de
codificacion.
}
\usage{codigos_territoriales}
\format{
Un data frame con 346 observaciones y seis columnas.
  \describe{
    \item{\code{codigo_comuna}}{vector de texto}
    \item{\code{nombre_comuna}}{vector de texto}
    \item{\code{codigo_provincia}}{vector de texto}
    \item{\code{nombre_provincia}}{vector de texto}
    \item{\code{codigo_region}}{vector de texto}
    \item{\code{nombre_region}}{vector de texto}
  }
}
\source{
Elaboracion propia con base en datos de Subsecretaria de Desarrollo Regional (SUBDERE).
}
\examples{
codigos_territoriales
}
\keyword{datasets}
